Refactor send transaction v2 (#6060)
* Refactor `eth_sendTransaction` method handler

The method handler for `eth_sendTransaction` has been moved to a
separate module to simplify unit testing. Unit tests have been added
as well.

This module was written in TypeScript, which asked for some additional
validation of the method parameters. We now throw a more explicit
message when the params are missing, and when the transaction
parameters are not an object. Previously the method would still throw
an error in these scenarios, just with a less helpful message. The
change in error message should be the only functional change here.

This refactor is intended to make testing PR #5619 easier.

* Support `.cjs` file extensions

The Metro configuration has been updated to support `.cjs` file
extensions. The `@metamask/utils` dependency relies upon the library
`superstruct` which uses `.cjs` for its main exported module.

* Move utils dependency to 'dependencies'

* Inline functions from `@metamask/utils`

The two functions taken from `@metamask/utils` have been inlined as a
workaround to an error that gets thrown upon import on Android builds.

I'm still actively investigating that failure, so hopefully we can
replace these inlined functions soon.